add new student feature and display them on a table added

+++ database.sql
@@ -30,13 +30,14 @@ CREATE TABLE "student" (
     "grad_year" DATE,
     "school_attend" varchar,
     "student_email" varchar NOT NULL,
-    "pif_amount" int,
+    "pif_amount" decimal,
     "password" VARCHAR (1000) NOT NULL,
     "created_at" DATE DEFAULT NOW(), --set default to today's date
     "lcf_start_date" DATE,
     "role" varchar,
     "last_login" DATE
 );
+
 
 CREATE TABLE "entry" (
   "id" SERIAL PRIMARY KEY,
